Changed port from long to int Implemented connect for tcp client

+++ src/SimpleNetworking/Networking/ITcpNetworkTransport.cs
@@ -7,6 +7,6 @@ namespace SimpleNetworking.Networking
 {
     public interface ITcpNetworkTransport 
     {
-        void Connect(string hostName, long port);
+        void Connect(string hostName, int port);
     }
 }
